added settings profile tabs
