+++ src/pages/profile/settings/settingsTabs/adminTab.tsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+interface Props{
+
+
+}
+
+
+function AdminSettingsTab(props:Props){
+
+
+    return(
+        <>
+            Личные настройки
+
+        </>
+    )
+}
+
+
+export default AdminSettingsTab;
