+++ src/pages/profile/settings/settingsTabs/appearanceTab.tsx
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+interface Props{
+
+
+}
+
+
+function AppearanceSettingsTab(props:Props){
+
+
+    return(
+        <>
+            Настройка окружения
+        </>
+    )
+}
+
+
+export default AppearanceSettingsTab;
